client is a reference to the class not a string
