+++ dataservice/__init__.py
@@ -1,5 +1,5 @@
 from dataservice.service import DataService
-from dataservice.ab_client import ABClient
+from dataservice.models import ABClient
 from dataservice.clients import HttpXClient
 from dataservice.models import Request, Response
 from dataservice.pipeline import Pipeline
