+++ dataservice/models.py
@@ -1,3 +1,6 @@
+from __future__ import annotations
+from abc import ABC, abstractmethod
+from logging import getLogger
 from typing import (
     Annotated,
     AsyncGenerator,
@@ -13,6 +16,7 @@ from typing import (
 
 from bs4 import BeautifulSoup
 from pydantic import AfterValidator, BaseModel, HttpUrl, model_validator
+
 
 DataItemGeneric = TypeVar("DataItemGeneric")
 RequestOrData = Union["Request", DataItemGeneric]
@@ -41,7 +45,7 @@ class Request(BaseModel):
     params: Optional[dict] = None
     form_data: Optional[dict] = None
     json_data: Optional[dict] = None
-    client: str
+    client: type(ABClient)
 
     @model_validator(mode="after")
     def validate(self):
@@ -72,3 +76,17 @@ class Response(BaseModel):
         if self.__soup is None:
             self.__soup = self.__get_soup()
         return self.__soup
+
+
+class ABClient(ABC):
+    """Abstract base class for clients."""
+
+    def __init__(self):
+        self.logger = getLogger(__name__)
+
+    def get_name(self):
+        return self.__class__.__name__
+
+    @abstractmethod
+    async def make_request(self, request: Request) -> Response:
+        raise NotImplementedError
